Rename agent_request to agent_message and move packages for websocket classes
