+++ src/main/java/io/quarkus/sample/agents/webstocket/AgentMessage.java
@@ -1,4 +1,4 @@
-package io.quarkus.sample.agent;
+package io.quarkus.sample.agents.webstocket;
 
 public record AgentMessage(Kind kind, String todoId, String payload) {
