+++ src/main/java/io/quarkus/sample/agents/webstocket/Kind.java
@@ -1,0 +1,5 @@
+package io.quarkus.sample.agents.webstocket;
+
+public enum Kind {
+    cancel, activity_log, agent_message, user_message
+}
